% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{assertVolcanoParams}
\alias{assertVolcanoParams}
\title{Assert volcano params}
\usage{
assertVolcanoParams(x, dt = NULL, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{dt}{Extra context for the check.}

\item{.var.name}{Name of the checked object to print in assertions.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns the checked object if the assertion is
successful.
}
\description{
Assert volcano params
}
\keyword{internal}
