% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_point_size}
\alias{auto_point_size}
\title{Automatic point size}
\usage{
auto_point_size(n_samples, raster = NULL)
}
\arguments{
\item{n_samples}{Integer. Number of samples.}

\item{raster}{Optional boolean. If rastering is activating.}
}
\value{
Automated dot size
}
\description{
Ported over from Seurats \code{AutoPointSize()}.
}
\keyword{internal}
