% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_ext.R
\name{checkBlitzGseaNull}
\alias{checkBlitzGseaNull}
\title{Check a blitzGSEA null model}
\usage{
checkBlitzGseaNull(x)
}
\arguments{
\item{x}{The object to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking a \code{BlitzGseaNull} object, i.e.
the output of \code{\link[bixverse:blitzgsea_calibrate]{bixverse::blitzgsea_calibrate()}}.
}
\keyword{internal}
