% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkVolcanoParams}
\alias{checkVolcanoParams}
\title{Check volcano params}
\usage{
checkVolcanoParams(x, dt = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{dt}{Extra context for the check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_volcano]{params_volcano()}}.
}
\keyword{internal}
