% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\docType{data}
\name{.colours_continuous_sc}
\alias{.colours_continuous_sc}
\title{Colour vector for continuous single-cell scales}
\format{
An object of class \code{character} of length 6.
}
\usage{
.colours_continuous_sc
}
\description{
Runs from light grey into the blue shades of MetBrewer
"Cassatt1".
}
\keyword{internal}
