% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{.palette_austria}
\alias{.palette_austria}
\title{Single-cell colour palette function for discrete scales}
\usage{
.palette_austria(n)
}
\arguments{
\item{n}{Integer. Number of groups found.}
}
\description{
Uses the MetBrewer "Austria" palette, auto-detecting the
required number of colours from the data.
}
\keyword{internal}
