% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse_plots.R
\name{enrichment_map_gsea}
\alias{enrichment_map_gsea}
\title{Generate enrichment map igraph (for GSEA)}
\usage{
enrichment_map_gsea(
  res,
  threshold,
  pathways,
  overlap_coefficient = FALSE,
  min_sim = 0.2,
  resolution = 1,
  layout_func = igraph::layout_with_fr,
  ...
)
}
\arguments{
\item{res}{data.table with the enrichment results. Needs to have the columns
\code{c("pathway_name", "nes", "fdr")}, i.e. the output of
\code{\link[bixverse:calc_fgsea]{bixverse::calc_fgsea()}} or \code{\link[bixverse:calc_blitzgsea]{bixverse::calc_blitzgsea()}}. A \code{geneset_name}
column is accepted with a deprecation warning.}

\item{threshold}{Numeric. The FDR threshold you wish to filter for.}

\item{pathways}{Named list. The original pathway list used for the
calculation of the overenrichment analysis.}

\item{overlap_coefficient}{Boolean. Shall the overlap coefficient be used
instead of the Jaccard similarity.}

\item{min_sim}{Numeric. Minimum similarity between two gene sets to be
connected.}

\item{resolution}{Numeric. The resolution parameter for the Louvain
clustering.}

\item{layout_func}{Layout function. Please see \code{\link[igraph:add_layout_]{igraph::add_layout_()}}
for options. This one will be used to layout the graph.}

\item{...}{Further parameters to forward to
\code{\link[=wrap_and_truncate]{wrap_and_truncate()}}.}
}
\value{
\code{igraph} object representing the enrichment map.
}
\description{
Helper function to generate an enrichment map based on GSEA results. Similar
enriched gene sets are clustered together via their Jaccard similarity
(alternatively overlap coefficient) and the function returns an igraph object
for subsequent visualisations.
}
