% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescence}
\usage{
a \%||\% b
}
\arguments{
\item{a}{R object a}

\item{b}{R object b}
}
\value{
If \code{a} is not \code{NULL}, a; otherwise b.
}
\description{
Null coalescence
}
\keyword{internal}
