% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qc_plots.R
\name{joint_plot}
\alias{joint_plot}
\title{Generic joint plot function}
\usage{
joint_plot(x, ...)
}
\arguments{
\item{x}{An object to plot.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list of ggplot objects.
}
\description{
Generic joint plot function
}
