% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qc_plots.R
\name{joint_plot_sc.CellQc}
\alias{joint_plot_sc.CellQc}
\title{Joint QC plot from a CellQc object}
\usage{
\method{joint_plot_sc}{CellQc}(
  x,
  library_size = "log10_lib_size",
  nb_features = "log10_nnz",
  palette = c("sequential", "spectral", "viridis", "diverging"),
  ...
)
}
\arguments{
\item{x}{A \code{CellQc} object.}

\item{library_size}{Character. Column with the library size per cell.}

\item{nb_features}{Character. Column with the number of features per cell.}

\item{palette}{String. Continuous palette for the hexbin fill. One of
\code{c("sequential", "spectral", "viridis", "diverging")}, see \code{\link[=bx_colors]{bx_colors()}}.}

\item{...}{Ignored.}
}
\value{
A \code{ggExtraPlot} object.
}
\description{
Joint QC plot from a CellQc object
}
