% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{joint_plot_sc}
\alias{joint_plot_sc}
\title{Generic joint plot function}
\usage{
joint_plot_sc(x, ...)
}
\arguments{
\item{x}{An object to plot.}

\item{...}{Arguments passed to the dispatched method.}
}
\value{
A \code{ggExtraPlot} object.
}
\description{
Dispatches to a joint hexbin plot of library size vs feature counts. See
\code{\link{joint_plot_sc.CellQc}} and
\code{\link{joint_plot_sc.data.table}} for the available methods.
}
\keyword{internal}
