% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_plots}
\alias{params_plots}
\title{Wrapper function for standard plot parameters}
\usage{
params_plots(
  width = 5,
  height = 5,
  file_type = c(".png", ".pdf"),
  unit = c("in", "px", "cm"),
  res = 450L,
  create_dir = TRUE
)
}
\arguments{
\item{width}{Numeric. Width of the plot. Defaults to \code{5.0}.}

\item{height}{Numeric. Height of the plot. Defaults to \code{5.0}.}

\item{file_type}{String. Plot type to save. Might be expanded to other file
types. One of \code{c(".png", ".pdf")}. Defaults to \code{".png"}.}

\item{unit}{String. Unit type for \code{width} and \code{height}. One of \code{c("in", "px", "cm")}. Defaults to \code{"in"}.}

\item{res}{Integer. Resolution for PNGs. Defaults to \code{450L}.}

\item{create_dir}{Boolean. Shall the plot directory be generated recursively.
Defaults to \code{TRUE}.}
}
\value{
A named list with the following elements:
\itemize{
\item width - Numeric. Width of the plot. Defaults to \code{5.0}.
\item height - Numeric. Height of the plot. Defaults to \code{5.0}.
\item file_type - String. Plot type to save. Might be expanded to other file
types. One of \code{c(".png", ".pdf")}. Defaults to \code{".png"}.
\item unit - String. Unit type for \code{width} and \code{height}. One of \code{c("in", "px", "cm")}. Defaults to \code{"in"}.
\item res - Integer. Resolution for PNGs. Defaults to \code{450L}.
\item create_dir - Boolean. Shall the plot directory be generated
recursively. Defaults to \code{TRUE}.
}
}
\description{
Wrapper function for standard plot parameters
}
