% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse_plots.R
\name{plot_enrichment_map_visnetwork}
\alias{plot_enrichment_map_visnetwork}
\title{igraph enrichment map to VisNetwork interactive network}
\usage{
plot_enrichment_map_visnetwork(g)
}
\arguments{
\item{g}{igraph. Output from \code{\link[=enrichment_map_oae]{enrichment_map_oae()}}.}
}
\value{
The interactive visnetwork
}
\description{
Takes in the output from \code{\link[=enrichment_map_oae]{enrichment_map_oae()}} and generates
an interactive VisNetwork widget.
}
