% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitch_plots.R
\name{prepare_mitch_scores}
\alias{prepare_mitch_scores}
\title{Generate adjusted scores for plotting mitch results}
\usage{
prepare_mitch_scores(res, fdr_threshold = 0.05)
}
\arguments{
\item{res}{data.table. Output of \code{\link[bixverse:calc_mitch]{bixverse::calc_mitch()}}.}

\item{fdr_threshold}{Numeric. The FDR threshold you want to apply.}
}
\value{
A list with the following two elements
\itemize{
\item fdr_corrections - FDR-corrected scores as a matrix of the individual
contrast p-values.
\item adj_scores - The extract scores per contrast with scores above the
individual FDR threshold set to 0.
}
}
\description{
Function that takes results from \code{\link[bixverse:calc_mitch]{bixverse::calc_mitch()}}, extracts the
individual scores and p-values for each contrast, runs an FDR correction
on top of the p-values and sets scores above the provided FDR threshold to
0.
}
