% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_helpers.R
\name{save_plot}
\alias{save_plot}
\title{General helper to save all types of plots to file}
\usage{
save_plot(plot, file_name = NULL, path, plot_params = params_plots())
}
\arguments{
\item{plot}{The plot you want to save to disk.}

\item{file_name}{Optional String. If not provided, the plot will be named
after the R variable.}

\item{path}{Directory. Where to save the plot to.}

\item{plot_params}{List. Output of \code{\link[=params_plots]{params_plots()}}. A list
with the following elements:
\itemize{
\item width - Width of the plot.
\item height - Height of the plot.
\item file_type - Which file type.
\item unit - Which unit do width and height describe.
\item res - Resolution for PNG plots
}}
}
\value{
Saves the plot to disk and returns \code{invisible}.
}
\description{
General helper to save all types of plots to file
}
