% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_helpers.R
\name{save_plot_ls}
\alias{save_plot_ls}
\title{Save a list of plots}
\usage{
save_plot_ls(plot_ls, path, plot_params = params_plots())
}
\arguments{
\item{plot_ls}{Named list of plots. The names will serve as file names (the
extension will be added by the function)}

\item{path}{Directory. Where to save the plot to.}

\item{plot_params}{List. Output of \code{\link[=params_plots]{params_plots()}}. A list
with the following elements:
\itemize{
\item width - Width of the plot.
\item height - Height of the plot.
\item file_type - Which file type.
\item unit - Which unit do width and height describe.
\item res - Resolution for PNG plots
}}
}
\value{
Saves the plots to disk and returns \code{invisible}.
}
\description{
Save a list of plots
}
