% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_helpers.R
\name{save_png}
\alias{save_png}
\title{Helper function to save various plot types to PNG}
\usage{
save_png(plot, f_path, plot_params)
}
\arguments{
\item{plot}{The plot you want to save to disk.}

\item{f_path}{File path. The path to which you want to save the PNG}

\item{plot_params}{List. Output of \code{\link[=params_plots]{params_plots()}}.}
}
\value{
Saves the PNG to disk.
}
\description{
Helper function to save various plot types to PNG
}
\keyword{internal}
