% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_bx}
\alias{scale_color_bx}
\title{Bixverse Color Scale (Discrete)}
\usage{
scale_color_bx(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name (default: "main")}

\item{reverse}{Reverse colors? (default: FALSE)}

\item{...}{Additional arguments passed to scale_color_manual}
}
\value{
A ggplot2 scale object
}
\description{
Bixverse Color Scale (Discrete)
}
