% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_bx}
\alias{scale_fill_bx}
\title{Bixverse Fill Scale (Discrete)}
\usage{
scale_fill_bx(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name (default: "main")}

\item{reverse}{Reverse colors? (default: FALSE)}

\item{...}{Additional arguments passed to scale_fill_manual}
}
\value{
A ggplot2 scale object
}
\description{
Bixverse Fill Scale (Discrete)
}
