% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plots.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Create a volcano plot}
\usage{
volcano_plot(
  dt,
  volcano_plot_params = params_volcano(),
  x_lab = NULL,
  y_lab = NULL,
  plot_title = NULL,
  plot_sub_title = NULL
)
}
\arguments{
\item{dt}{A data.table/data.frame holding the differential expression results.}

\item{volcano_plot_params}{A list as returned by \code{\link[=params_volcano]{params_volcano()}}.}

\item{x_lab}{Optional string. Overwrite of the x label of the plot. If not
provided, will default to \code{volcano_plot_params$x_axis}.}

\item{y_lab}{Optional string. Overwrite of the x label of the plot. If not
provided, will default to \code{-log10(volcano_plot_params$y_axis)}.}

\item{plot_title}{Optional string. If provided, the plot will get a title.}

\item{plot_sub_title}{Optional string. If provided, adds this as a sub title
to the plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Helper function to generate volcano plots for differental
gene/transcript/protein expression analysis. The function applies under the
hood a \code{-log10(stat)} transformation to the respective statistic you want
to display on the y axis. If you provide labels, the top features to label
will be chosen by ranking the \code{-log10(stat)}.
}
