% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_and_truncate}
\alias{wrap_and_truncate}
\title{Helper function to wrap and truncate text}
\usage{
wrap_and_truncate(text, width = 40L, max_lines = 2L, whitespace_only = TRUE)
}
\arguments{
\item{text}{String. The string to truncate.}

\item{width}{Integer. Maximum width of a given line.}

\item{max_lines}{Integer. Maximum lines before truncating.}

\item{whitespace_only}{Boolean. Shall the string wrapping happen only around
whitespaces.}
}
\value{
The string wrapped and/or truncated.
}
\description{
Helper function to wrap and truncate text
}
