% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{bx_colors}
\alias{bx_colors}
\title{Bixverse Color Palette}
\usage{
bx_colors(palette = "main", reverse = FALSE, n = 20, ...)
}
\arguments{
\item{palette}{Palette name. One of \code{"main"}, \code{"sequential"}, \code{"diverging"},
\code{"viridis"} or \code{"spectral"}. The latter is a reversed RColorBrewer Spectral,
i.e. dark indigo for low and dark red for high values.}

\item{reverse}{Logical, reverse the color order? (default: FALSE)}

\item{n}{Integer, number of colours to return.}

\item{...}{Ignored. Present for backwards compatibility.}
}
\value{
A vector of color hex codes
}
\description{
Official bixverse color palette
}
\keyword{internal}
