% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{bx_colors}
\alias{bx_colors}
\title{Bixverse Color Palette}
\usage{
bx_colors(palette = "main", reverse = FALSE)
}
\arguments{
\item{palette}{Palette name: "main", "diverging", "sequential"}

\item{reverse}{Logical, reverse the color order? (default: FALSE)}
}
\value{
A vector of color hex codes
}
\description{
Official bixverse color palette
}
\keyword{internal}
