% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_ext.R
\name{checkPlotParams}
\alias{checkPlotParams}
\title{Check general graph parameters}
\usage{
checkPlotParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the general plot parameters.
}
