% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qc_plots.R
\name{density_plot_sc.data.table}
\alias{density_plot_sc.data.table}
\title{Plot per-cell QC density plots from a data.table}
\usage{
\method{density_plot_sc}{data.table}(
  df,
  grouping_column,
  variable,
  var_name = NULL,
  nmads = 3,
  log_scale = TRUE,
  adjust_position_label = 0
)
}
\arguments{
\item{df}{data.table. Input data table containing QC metrics.}

\item{grouping_column}{Character. Name of the column used to group density plot}

\item{variable}{Character. Name of the numeric QC column to plot on the x-axis.}

\item{var_name}{Character. Label for the y-axis (default: NULL).}

\item{log_scale}{Boolean. Log scale (default: TRUE)}

\item{group_name}{Character. Label for the x-axis (default: NULL).}

\item{show_outlier}{Boolean. Show the cells that are identified as outliers (default: TRUE)}

\item{outlier_column}{Character. Which column contains the cells identified as outliers}
}
\value{
A named list of ggplot objects, one per metric.
}
\description{
Plot per-cell QC density plots from a data.table
}
\keyword{internal}
