% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qc_plots.R
\name{density_plot_sc}
\alias{density_plot_sc}
\title{Generic density plot function}
\usage{
density_plot_sc(x, ...)
}
\arguments{
\item{x}{An object to plot.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list of ggplot objects.
}
\description{
Generic density plot function
}
