% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{.colours_continuous_sc}
\alias{.colours_continuous_sc}
\title{Colour vector for continuous single-cell scales}
\usage{
.colours_continuous_sc
}
\description{
Runs from light grey into the blue shades of MetBrewer
"Cassatt1".
}
\keyword{internal}
