% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qc_plots.R
\name{.plot_joint}
\alias{.plot_joint}
\title{Joint hexbin plot worker}
\usage{
.plot_joint(
  df,
  library_size = "lib_size",
  nb_features = "nnz",
  log_scale = FALSE
)
}
\arguments{
\item{df}{data.table. Plotting-ready data.}

\item{library_size}{Character. Column with the library size per cell.}

\item{nb_features}{Character. Column with the number of features per cell.}

\item{log_scale}{Logical. Log10-transform both axes (default: FALSE).}
}
\value{
A \code{ggExtraPlot} object.
}
\description{
Joint hexbin plot worker
}
\keyword{internal}
