% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_embd_dot_vln_plots.R
\name{dot_plot_sc}
\alias{dot_plot_sc}
\title{Dot plot of marker gene expression across groups}
\usage{
dot_plot_sc(
  object,
  features,
  grouping_variable,
  feature_labels = NULL,
  cell_markers = NULL,
  scale_exp = TRUE,
  modality = c("rna", "adt")
)
}
\arguments{
\item{object}{A single cell class.}

\item{features}{Character vector. Gene IDs to plot.}

\item{grouping_variable}{String. Obs column to group by.}

\item{feature_labels}{Optional named character vector mapping gene ids to
display labels (default: NULL).}

\item{cell_markers}{Optional named character vector mapping gene ids to
cell type labels. If feature_labels is provided, the character vectors
should contain the mapping of feature display labels to cell type. E.g.
c(CD3E = "T cell", CD8A = "T cell", MS4A1 = "B cell", ...).
(default: NULL).}

\item{scale_exp}{Boolean. Whether to min-max scale mean expression per gene.}

\item{modality}{String. One of \code{c("rna", "adt")}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Dot plot of marker gene expression across groups
}
