% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_plots.R
\name{dot_plot_sc}
\alias{dot_plot_sc}
\title{Dot plot of marker gene expression across groups}
\usage{
dot_plot_sc(
  object,
  features,
  grouping_variable,
  feature_labels = NULL,
  feature_grouping = NULL,
  scale_exp = TRUE,
  modality = c("rna", "adt"),
  cluster_groups = TRUE
)
}
\arguments{
\item{object}{A single cell class.}

\item{features}{Character vector. Gene IDs to plot.}

\item{grouping_variable}{String. Obs column to group by.}

\item{feature_labels}{Optional named character vector mapping gene ids to
display labels (default: NULL).}

\item{feature_grouping}{Optional named character vector mapping gene ids to
grouping labels, e.g. cell type labels. If feature_labels is provided,
the character vectors should contain the mapping of feature display labels to
their respecitve groups (e.g. c(CD3E = "T cell", CD8A = "T cell",
MS4A1 = "B cell", ...). (default: NULL).}

\item{scale_exp}{Boolean. Whether to min-max scale mean expression per gene.}

\item{modality}{String. One of \code{c("rna", "adt")}.}

\item{cluster_groups}{Boolean. Use hierarchical clustering on the grouping variable
to re-order the group labels based on expression similarity.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Dot plot of marker gene expression across groups
}
