% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_embd_dot_vln_plots.R
\name{feature_plot_sc}
\alias{feature_plot_sc}
\title{Faceted feature plot over an embedding}
\usage{
feature_plot_sc(
  object,
  features,
  embedding,
  feature_labels = NULL,
  scale = FALSE,
  clip = NULL,
  modality = c("rna", "adt"),
  point_size = NULL,
  raster = NULL,
  raster_dpi = c(512, 512),
  highlight_features = FALSE,
  highlight_quantile = 0.25,
  ...
)
}
\arguments{
\item{object}{A single cell class.}

\item{features}{Character vector. Gene IDs to plot.}

\item{embedding}{String. Name of the embedding.}

\item{feature_labels}{Optional named character vector mapping gene ids to
display labels (default: NULL).}

\item{scale}{Boolean. Whether to z-score the expression values.}

\item{clip}{Optional numeric. Clip z-scores if \code{scale = TRUE}.}

\item{modality}{String. One of \code{c("rna", "adt")}.}

\item{point_size}{Optional numeric. Defines the point size. If not provided,
will be auto-determined.}

\item{raster}{Optional boolean. Shall the plot be rasterised. If \code{NULL} and
number of cells is larger than \code{1e5}, defaults to TRUE.}

\item{raster_dpi}{Two numerics. Pixel resolution for rasterized plots, passed
to geom_scattermore(). Default is \code{c(512, 512)}.}

\item{highlight_features}{Boolean. Shall the features be more strongly
highlighted. Useful for sparsely expressed genes.}

\item{highlight_quantile}{Numeric between \verb{[0, 1]}. Defines the threshold.}

\item{...}{Additional arguments forwarded to
\code{\link[bixverse:extract_embedding_data]{bixverse::extract_embedding_data()}}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Faceted feature plot over an embedding
}
