% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plots.R
\name{get_gsea_enrichment_data}
\alias{get_gsea_enrichment_data}
\title{Helper function to get the plot data for GSEA plots}
\usage{
get_gsea_enrichment_data(
  stats,
  pathways,
  pathways_of_interest,
  gsea_results = NULL,
  gsea_param = 1
)
}
\arguments{
\item{stats}{Named numeric vector. The gene level statistic.}

\item{pathways}{List. A named list with each element containing the genes for
this pathway.}

\item{pathways_of_interest}{String vector. Names of the pathways to plot.
These strings need to be represented in the names of pathways.}

\item{gsea_results}{Optional data.table with the bixverse GSEA results. If
provided, the FDR and NES for the given pathway of interest will be also
added to the plot.}

\item{gsea_param}{Numeric. Defaults to \code{1}}
}
\value{
A list of \code{gsea_par_plot_data}
}
\description{
Helper function to get the plot data for GSEA plots
}
