% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse_plots.R
\name{helper_gse_dot_plot}
\alias{helper_gse_dot_plot}
\title{Helper to generate a GSE dot plot}
\usage{
helper_gse_dot_plot(
  res,
  size_range = c(2, 5),
  viridis_option = "D",
  direction = -1
)
}
\arguments{
\item{res}{data.table with the enrichment results. Needs to have the columns
\code{c("hits", "target_set_lengths", "gene_set_name", "gene_set_lengths", "fdr")}.}

\item{size_range}{Numerical vector of size 2. Defines the size range for the
dots in the plot.}

\item{viridis_option}{String. The option to forward to
\code{\link[ggplot2:scale_fill_viridis_c]{ggplot2::scale_fill_viridis_c()}}.}

\item{direction}{\code{1} or \code{-1}. The direction in the colour palette.}
}
\value{
The GSE dot plot.
}
\description{
Helper to generate a GSE dot plot
}
\keyword{internal}
