% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qc_plots.R
\name{joint_plot_sc.CellQc}
\alias{joint_plot_sc.CellQc}
\title{Joint QC plots from a CellQc object}
\usage{
\method{joint_plot_sc}{CellQc}(x, ...)
}
\arguments{
\item{x}{A \code{CellQc} object.}

\item{...}{Ignored.}
}
\value{
A named list of ggplot objects, one per metric.
}
\description{
Joint QC plots from a CellQc object
}
