% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qc_plots.R
\name{joint_plot_sc.data.table}
\alias{joint_plot_sc.data.table}
\title{Joint QC plot from a data.table}
\usage{
\method{joint_plot_sc}{data.table}(
  x,
  library_size = "lib_size",
  nb_features = "nnz",
  log_scale = FALSE,
  ...
)
}
\arguments{
\item{x}{data.table. Input data containing the QC metrics.}

\item{library_size}{Character. Column with the library size per cell.}

\item{nb_features}{Character. Column with the number of features per cell.}

\item{log_scale}{Logical. Log10-transform both axes (default: FALSE).}

\item{...}{Ignored.}
}
\value{
A \code{ggExtraPlot} object.
}
\description{
Joint QC plot from a data.table
}
