% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_wrappers.R
\name{params_plots}
\alias{params_plots}
\title{Wrapper function for standard plot parameters}
\usage{
params_plots(
  width = 5,
  height = 5,
  file_type = c(".png", ".pdf"),
  unit = c("in", "px", "cm"),
  res = 450L,
  create_dir = TRUE
)
}
\arguments{
\item{width}{Float. Width of the plot.}

\item{height}{Float. Height of the plot.}

\item{file_type}{String. One of \code{c(".png", "pdf")}. Plot type to save. Might
be expanded to other file types. Defaults to \code{".png"}}

\item{unit}{String. One of \code{c("in", "px", "cm")}. Unit type for \code{width} and
\code{height}. Defaults to \code{"in"}.}

\item{res}{Integer. Resolution for PNGs.}

\item{create_dir}{Boolean. Shall the plot directory be generated recursively.}
}
\value{
A list with the parameters for usage in subsequent functions.
}
\description{
Wrapper function for standard plot parameters
}
