% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse_plots.R
\name{plot_enrichment_map_ggraph}
\alias{plot_enrichment_map_ggraph}
\title{igraph enrichment map to ggraph plot}
\usage{
plot_enrichment_map_ggraph(
  g,
  label_nodes = "adaptive",
  adaptive_thresholds = c(`15` = 3, `5` = 2, `2` = 1, `1` = 0),
  font_size = 4
)
}
\arguments{
\item{g}{igraph. Output from \code{\link[=enrichment_map_oae]{enrichment_map_oae()}}.}

\item{label_nodes}{String. Controls which nodes to label. Options:
\itemize{
\item \code{"all"}: Label all nodes
\item \code{"adaptive"}: Adaptive labelling based on community size (default)
\item \code{NULL}: No labels
\item Integer: Label top N nodes by size
}}

\item{adaptive_thresholds}{Named numeric. The names indicate the community
size and the values how many pathways per community to show. An example would
be \verb{c(15 = 3, 5 = 2, 2 = 1, 1 = 0)}}

\item{font_size}{Numeric. Font size of the labels on top of the enrichment
map.}
}
\value{
A ggplot2 object with the enrichment map
}
\description{
Takes in the output from \code{\link[=enrichment_map_oae]{enrichment_map_oae()}} and generates
a ggraph object for subsequent saving, etc.
}
