% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse_plots.R
\name{plot_gse_dotplot}
\alias{plot_gse_dotplot}
\title{Generate GSE dotplots}
\usage{
plot_gse_dotplot(
  res,
  size_range = c(2, 5),
  viridis_option = "D",
  direction = -1,
  .verbose = TRUE
)
}
\arguments{
\item{res}{data.table with the enrichment results. Needs to have the columns
\code{c("hits", "target_set_lengths", "gene_set_name", "gene_set_lengths", "fdr")}.}

\item{size_range}{Numerical vector of size 2. Defines the size range for the
dots in the plot.}

\item{viridis_option}{String. The option to forward to
\code{\link[ggplot2:scale_fill_viridis_c]{ggplot2::scale_fill_viridis_c()}}.}

\item{direction}{\code{1} or \code{-1}. The direction in the colour palette.}

\item{.verbose}{Boolean. Controls verbosity of the function.}
}
\value{
If the output of \code{\link[bixverse:gse_hypergeometric_list]{bixverse::gse_hypergeometric_list()}} was provided,
a list of dotplots per target gene set. Otherwise, a single GSE OAE dot plot.
}
\description{
This function can take in the output of \code{\link[bixverse:gse_hypergeometric]{bixverse::gse_hypergeometric()}} or
\code{\link[bixverse:gse_hypergeometric_list]{bixverse::gse_hypergeometric_list()}} and generates in the former case a
single plot and in the latter case a list of plots per target set.
}
