% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plots.R
\name{plot_gsea_enrichment}
\alias{plot_gsea_enrichment}
\title{Plot GSEA enrichment results}
\usage{
plot_gsea_enrichment(
  stats,
  pathways,
  pathways_of_interest,
  gsea_results = NULL,
  gsea_param = 1,
  tick_size = 0.2,
  text_size = 5
)
}
\arguments{
\item{stats}{Named numeric vector. The gene level statistic.}

\item{pathways}{List. A named list with each element containing the genes for
this pathway.}

\item{pathways_of_interest}{String vector. Names of the pathways to plot.
These strings need to be represented in the names of pathways.}

\item{gsea_results}{Optional data.table with the bixverse GSEA results. If
provided, the FDR and NES for the given pathway of interest will be also
added to the plot.}

\item{gsea_param}{Numeric. Defaults to \code{1}.}

\item{tick_size}{Numeric. The tick size. Defaults ot \code{0.2}.}

\item{text_size}{Numeric. The text size. Defaults ot \code{8}. Only relevant when
\code{gsea_results} is provided.}
}
\description{
Helper function to create the classical GSEA plots for a set of pathways
of interest. You can also provide the underlying GSEA results to add the
FDR and NES information to the plot.
}
