% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_bx_c}
\alias{scale_color_bx_c}
\title{Bixverse Color Scale (Continuous)}
\usage{
scale_color_bx_c(palette = "sequential", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name (default: "sequential")}

\item{reverse}{Reverse colors? (default: FALSE)}

\item{...}{Additional arguments passed to scale_color_gradientn}
}
\value{
A ggplot2 scale object
}
\description{
Bixverse Color Scale (Continuous)
}
