% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_bx_c}
\alias{scale_fill_bx_c}
\title{Bixverse Fill Scale (Continuous)}
\usage{
scale_fill_bx_c(palette = "sequential", reverse = FALSE, n = 20, ...)
}
\arguments{
\item{palette}{Palette name (default: "sequential")}

\item{reverse}{Reverse colors? (default: FALSE)}

\item{n}{Integer. Number of colours to interpolate over (default: 20). Only
has an effect on palettes that respect \code{n}, see \code{\link[=bx_colors]{bx_colors()}}.}

\item{...}{Additional arguments passed to scale_fill_gradientn}
}
\value{
A ggplot2 scale object
}
\description{
Bixverse Fill Scale (Continuous)
}
