% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{set_bx_theme}
\alias{set_bx_theme}
\title{Set Bixverse Theme as Default}
\usage{
set_bx_theme(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{Base font size (default: 12)}

\item{base_family}{Base font family (default: "Arial")}
}
\value{
NULL (sets theme invisibly)
}
\description{
Sets the Bixverse theme as the default for all ggplot2 plots in the session
}
\examples{
\dontrun{
set_bx_theme()
}

}
\keyword{internal}
