% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qc_plots.R
\name{violin_plot_sc.CellQc}
\alias{violin_plot_sc.CellQc}
\title{Plot per-cell QC violin plots from a CellQc object}
\usage{
\method{violin_plot_sc}{CellQc}(
  x,
  grouping_column = "grp",
  variable,
  group_name = NULL,
  var_name = NULL,
  log_scale = FALSE,
  outlier_column = "global_outlier",
  show_outlier = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{CellQc} object.}

\item{...}{Ignored.}
}
\value{
A named list of ggplot objects, one per metric.
}
\description{
Plot per-cell QC violin plots from a CellQc object
}
\keyword{internal}
