% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_qc_plots.R
\name{violin_plot_sc}
\alias{violin_plot_sc}
\title{Generic violin plot function}
\usage{
violin_plot_sc(x, ...)
}
\arguments{
\item{x}{An object to plot.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A named list of ggplot objects.
}
\description{
Generic violin plot function
}
